[AC-1081] Added tests
