+++ src/components/breadcrumbs/tests/BreadCrumbs.test.js
@@ -1,0 +1,21 @@
+import React from 'react';
+import { shallow } from 'enzyme';
+import { BreadCrumbs, BreadCrumbsItem } from '../BreadCrumbs';
+
+describe('BreadCrumbs', () => {
+
+  const defaultProps = {
+    children: [
+      <BreadCrumbsItem onClick={jest.fn()} active={false}>A</BreadCrumbsItem>,
+      <BreadCrumbsItem onClick={jest.fn()} active={false}>B</BreadCrumbsItem>,
+      <BreadCrumbsItem onClick={jest.fn()} active={false}>C</BreadCrumbsItem>
+    ]
+  };
+
+  const subject = (props) => (shallow(<BreadCrumbs {...defaultProps} {...props}/>));
+
+  it('should not render > after the last breadCrumb ' , () => {
+    const instance = subject();
+    expect(instance.find('div').children().last().text()).not.toBe('>');
+  });
+});
